Liquibase migration added

+++ src/main/java/ru/baskaeva/dancehallsteps/DTO/StepsDTO.java
@@ -11,6 +11,6 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class StepsDTO {
-    Long totalCount;
+    Integer totalCount;
     List<Step> steps;
 }
